CMake simplification
Allright, previously cmake was configured to generate all kinds of
rules for exporting. BUT this library is not meant to be installed
into any system. It is intended to be used as an embedded
source-library. That way it can be integrated into your own source as an
external component that uses all the same compile options as your parent
project does. This is similar to how googletest works.